Fix for #115 - only check the g-force and decouple if we're attached to a real vessel, and not when on EVA.
When we're removed from a container during EVA there seems to be a moment
of high g-force which triggers the decouple.  This creates a phantom
duplicate. This phantom duplicate then causes all sorts of problems
in KSP, including spamming the log with NullReferenceExceptions and being
unable to return to the KSC or abort the flight.